BAH-2129 | refactor. restructure directories
