+++ src/common/userHasAccess/UserHasAccess.component.tsx
