+++ src/home-button/home-button.tsx
@@ -1,7 +1,7 @@
 import Home24 from '@carbon/icons-react/lib/home/24'
 import {HeaderGlobalAction} from 'carbon-components-react'
 import React from 'react'
-import {bahmniHomePath} from '../constants'
+import {bahmniHomePath} from '../utils/constants'
 import styles from './home-button.scss'
 
 const HomeButton = () => {
