+++ src/hooks/useOrderTypeUuidConfig.ts
@@ -1,7 +1,7 @@
 import {useEffect} from 'react'
 import useSWR from 'swr'
 import createPersistedState from 'use-persisted-state'
-import {orderTypeUuidKey, labOrderType} from '../constants'
+import {orderTypeUuidKey, labOrderType} from '../utils/constants'
 import {OrderTypeFetchResponse, OrderType} from '../types'
 import {getOrderTypeUuid} from '../utils/api-utils'
 import {fetcher} from '../utils/api-utils'
