+++ src/patient-lab-dashboard/image-preview-component/image-preview-component.scss
