+++ src/patient-lab-dashboard/patient-lab-details/patient-lab-details.scss
