+++ src/patient-lab-dashboard/table/report-table/report-table.component.scss
