+++ src/patient-lab-dashboard/upload-file/upload-file.scss
