+++ src/patient-lab-dashboard/upload-file/upload-file.test.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import {render, screen} from '@testing-library/react'
 import UploadFile from './upload-file'
 import userEvent from '@testing-library/user-event'
-import {UploadReportProvider} from '../context/upload-report-context'
-import {uploadFiles} from '../utils/test-utils/upload-report-helper'
+import {UploadReportProvider} from '../../context/upload-report-context'
+import {uploadFiles} from '../../utils/test-utils/upload-report-helper'
 
 describe('upload file', () => {
   it('should show the file upload box', () => {
